Ubuntu/18.04

+++ install.sh
@@ -1,37 +1,32 @@
 #!/usr/bin/env bash
 
-set -e
+set -o errexit
+set -o pipefail
+set -o nounset
+# set -o xtrace
 
-INSTALLER_VERSION="18.04"
+CONFIGURATOR_VERSION="18.04"
 
-UBUNTU_VERSION=`lsb_release -r -s`
-ROOT_DIR=$(dirname "$0")
-INSTALLER_DIR="${ROOT_DIR}/installer"
+ROOT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+MODULES_DIR="${ROOT_DIR}/modules"
+CONFIG_FILE="${ROOT_DIR}/config.sh"
 
-progress(){
-    notify-send "Installation progress: ${1}"
-}
-
-message(){
-    zenity --text "${1}" --info --width=200 --height=50 > /dev/null 2>&1
-}
-
-input(){
+input() {
     zenity --entry --title="Ubuntu Configurator" --text="${1}"
 }
 
-if [[ "${UBUNTU_VERSION}" != "${INSTALLER_VERSION}" ]] ; then
-    message "ERROR: This installer is only for Ubuntu "
-    exit 1;
-fi
+source ${MODULES_DIR}/check.sh
+
+sudo echo -e "\e[32mLet's start the installation ;)\e[39m";
 
 CHOICES=$(whiptail --checklist "Select which services do you want install. " \
     20 55 15 \
-    "zsh" "Z Shell" off \
     "diagnostic" "Diagnostic tools" on \
+    "zsh" "Z Shell" off \
     "google-chrome" "Google Chrome" on \
+    "gpg" "GPG Key" off \
     "git" "Git" on \
-    "git-hooks" "Hooks for Git" off \
+    "git-hooks" "Hooks for Git" on \
     "jetbrains-toolbox" "Jetbrains Toolbox" on \
     "docker" "Docker & Docker Compose" on  \
     "ssh-keygen" "Generate RSA key" on \
@@ -40,19 +35,31 @@ CHOICES=$(whiptail --checklist "Select which services do you want install. " \
     "ufw" "Enable firewall" on \
     "vpn-client" "VPN Client" on \
     "php" "PHP 7.2 with extensions" off  \
-    "yarn" "Node.js + Yarn" off \
+    "frontend-tools" "Node.js + Yarn + Vue" off \
     "spotify" "Spotify" off \
     "ssh-server" "Open SSH Server" off \
-    "gpg" "GPG Key" off \
-    "java8" "Java 8" off \
     "atom" "Atom editor" off \
+    "virtualbox" "VirtualBox" off \
+    "rest-tools" "REST Tools" off \
+    "sublime3" "Sublime Text 3" off \
     3>&2 2>&1 1>&3 )
 
-source ${INSTALLER_DIR}/misc.sh
+if [[ -f ${CONFIG_FILE} ]]; then
+    source ${CONFIG_FILE}
+fi
+
+source ${MODULES_DIR}/misc.sh
 
 for CHOICE in ${CHOICES}; do
-    source "${INSTALLER_DIR}/`echo ${CHOICE} | tr -d '"'`.sh"
+    source "${MODULES_DIR}/`echo ${CHOICE} | tr -d '"'`.sh"
 done
 
-message "Press [OK] to reboot system"
-reboot
+if [[ -f ${CONFIG_FILE} ]]; then
+    if zenity --question --text="Do you want safely remove your configuration?"; then
+        shred --remove --iterations=100 ${CONFIG_FILE}
+    fi
+fi
+
+if zenity --question --text="Do you want to reboot your system?"; then
+    reboot
+fi
